refactor: Route URL opening and file watching through the host
