+++ src/CutTheRopeDX.Desktop/Desktop/DesktopHostApp.cs
@@ -1,3 +1,7 @@
+using System;
+using System.ComponentModel;
+using System.Diagnostics;
+
 using CutTheRopeDX.Framework.Core;
 using CutTheRopeDX.Framework.Platform;
 
@@ -22,6 +26,27 @@ namespace CutTheRopeDX.Desktop
         public void DrawMovie()
         {
             Global.XnaGame.DrawMovie();
+        }
+
+        /// <inheritdoc />
+        public void OpenUrl(string url)
+        {
+            try
+            {
+                ProcessStartInfo psi = new()
+                {
+                    FileName = url,
+                    UseShellExecute = true
+                };
+                _ = Process.Start(psi);
+            }
+            catch (Win32Exception ex)
+            {
+                int errorCode = ex.ErrorCode;
+            }
+            catch (Exception)
+            {
+            }
         }
     }
 }
